Fix goal ID truncation and tighten server query/deps
Android was reading goal_id as int, silently truncating Postgres
BIGSERIAL IDs once they exceed 32 bits. Summary endpoint ran two
sequential queries per request; collapsed into one CTE. Pinned
Express to ^4.21.0 so async route errors reach the error handler.

+++ app/src/main/java/com/example/savingsgoal/Activities/AddEditGoalActivity.java
@@ -43,7 +43,7 @@ public class AddEditGoalActivity extends AppCompatActivity {
         repo = Repository.get(this);
         initViews();
 
-        goalId = getIntent().getIntExtra("goal_id", -1);
+        goalId = getIntent().getLongExtra("goal_id", -1);
 
         if (goalId != -1L) {
             toolbar.setTitle(R.string.title_edit_goal);
